#112 : More cleaning up strategy config loading
